requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langgraph
redis
faiss-cpu
google-genai

